requirements.txt
gd.py
python-dotenv
